move multiply decorator to a separate file
